+++ common/src/main/resources/sodium-common.mixins.json
@@ -67,7 +67,6 @@
     "features.render.immediate.buffer_builder.sorting.MultiBufferSourceMixin",
     "features.render.immediate.buffer_builder.sorting.VertexSortingMixin",
     "features.render.immediate.matrix_stack.VertexConsumerMixin",
-    "features.render.model.ItemModelGeneratorMixin",
     "features.render.particle.QuadParticleRenderStateMixin",
     "features.render.viewport.GlStateManagerMixin",
     "features.render.world.clouds.CloudRendererMixin",
