+++ fabric/src/main/resources/sodium-fabric.mixins.json
@@ -12,6 +12,7 @@
   "client": [
     "core.model.quad.BakedQuadMixin",
     "fabric.LevelSliceMixin",
+    "features.render.model.ItemModelGeneratorMixin",
     "features.world.biome.BiomeMixin"
   ]
 }
