+++ neoforge/src/mod/resources/META-INF/accesstransformer.cfg
@@ -1,24 +1,22 @@
 public net.minecraft.client.model.geom.ModelPart$Vertex
 public net.minecraft.client.model.geom.ModelPart$Polygon
-public net.minecraft.client.renderer.texture.SpriteContents$InterpolationData
 public net.minecraft.client.renderer.texture.SpriteContents$AnimatedTexture
+public net.minecraft.client.renderer.texture.SpriteContents$FrameInfo <init>(II)V
 public net.minecraft.client.renderer.texture.SpriteContents$FrameInfo
-public net.minecraft.client.renderer.texture.SpriteContents$Ticker
+public net.minecraft.client.renderer.texture.SpriteContents$AnimatedTexture interpolateFrames
+public net.minecraft.client.renderer.texture.SpriteContents$AnimatedTexture frames
+public net.minecraft.world.level.chunk.PalettedContainer$Data <init>(Lnet/minecraft/world/level/chunk/Configuration;Lnet/minecraft/util/BitStorage;Lnet/minecraft/world/level/chunk/Palette;)V
 public net.minecraft.world.level.chunk.PalettedContainer$Data
-public net.minecraft.world.level.chunk.PalettedContainer$Configuration
-public net.minecraft.client.renderer.FogRenderer$FogData
-public net.minecraft.client.renderer.FogRenderer$MobEffectFogFunction
+public net.minecraft.client.renderer.texture.Stitcher$Holder <init>(Lnet/minecraft/client/renderer/texture/Stitcher$Entry;II)V
 public net.minecraft.client.renderer.texture.Stitcher$Holder
 public net.minecraft.world.level.biome.Biome$ClimateSettings
 public net.minecraft.client.renderer.SectionBufferBuilderPool <init>(Ljava/util/List;)V
 
-public com.mojang.blaze3d.vertex.PoseStack$Pose <init>(Lorg/joml/Matrix4f;Lorg/joml/Matrix3f;)V
 public com.mojang.blaze3d.vertex.PoseStack$Pose trustedNormals # trustedNormals
 public com.mojang.blaze3d.opengl.GlDevice
 
 public net.minecraft.world.level.GrassColor pixels
 public net.minecraft.world.level.FoliageColor pixels
-public net.minecraft.client.renderer.FogRenderer getPriorityFogFunction(Lnet/minecraft/world/entity/Entity;F)Lnet/minecraft/client/renderer/FogRenderer$MobEffectFogFunction;
 
 public net.minecraft.client.resources.model.cuboid.ItemModelGenerator MIN_Z
 public net.minecraft.client.resources.model.cuboid.ItemModelGenerator MAX_Z
@@ -26,4 +24,3 @@ public net.minecraft.client.resources.model.cuboid.ItemModelGenerator UV_SHRINK
 public net.minecraft.client.resources.model.cuboid.ItemModelGenerator isTransparent(Lnet/minecraft/client/renderer/texture/SpriteContents;IIIII)Z
 public net.minecraft.client.resources.model.cuboid.ItemModelGenerator$SideDirection
 public net.minecraft.client.resources.model.cuboid.ItemModelGenerator$SideDirection isHorizontal()Z
-public net.minecraft.client.resources.model.cuboid.ItemModelGenerator$ItemLayerKey
