+++ neoforge/src/mod/resources/sodium-neoforge.mixins.json
@@ -11,6 +11,7 @@
   },
   "client": [
     "core.model.quad.BakedQuadMixin",
+    "features.render.model.ItemModelGeneratorMixin",
     "features.world.biome.BiomeMixin",
     "platform.neoforge.AbstractBlockRenderContextMixin",
     "platform.neoforge.AuxiliaryLightManagerMixin",
